Rename FolderPicker to OldFolderPicker
